pochti1

+++ app/src/main/java/com/example/krmobil/dbhelper/DBHelper.kt
@@ -9,37 +9,50 @@ import com.example.krmobil.models.Tool
 import com.example.krmobil.models.Material
 import com.example.krmobil.models.Sale
 import com.example.krmobil.models.User
+import com.example.krmobil.models.Comment
 import com.example.krmobil.utils.CryptoUtils
 import javax.crypto.SecretKey
 
 class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
-    SQLiteOpenHelper(context, "app", factory, 16) { // Увеличьте версию базы данных
+    SQLiteOpenHelper(context, "app", factory, 21) {
 
     private val secretKey: SecretKey = CryptoUtils.getKey(context)
 
-    // DBHelper.kt
     override fun onCreate(db: SQLiteDatabase?) {
         val queryTools = "CREATE TABLE tools (id INTEGER PRIMARY KEY AUTOINCREMENT, image TEXT, name TEXT, description TEXT, price REAL, category TEXT)"
         val queryMaterials = "CREATE TABLE materials (id INTEGER PRIMARY KEY AUTOINCREMENT, image TEXT, name TEXT, description TEXT, price REAL, category TEXT)"
         val queryUsers = "CREATE TABLE users (id INTEGER PRIMARY KEY AUTOINCREMENT, email TEXT, login TEXT, phone TEXT, pass TEXT, is_admin INTEGER)"
-        val querySales = "CREATE TABLE sales (id INTEGER PRIMARY KEY AUTOINCREMENT, item_id INTEGER, item_type TEXT, quantity INTEGER, sale_date TEXT, FOREIGN KEY(item_id) REFERENCES tools(id) ON DELETE CASCADE, FOREIGN KEY(item_id) REFERENCES materials(id) ON DELETE CASCADE)"
+        val querySales = "CREATE TABLE sales (id INTEGER PRIMARY KEY AUTOINCREMENT, item_id INTEGER, item_type TEXT, item_name TEXT, item_image TEXT, item_description TEXT, price REAL, quantity INTEGER, sale_date TEXT, FOREIGN KEY(item_id) REFERENCES tools(id) ON DELETE CASCADE, FOREIGN KEY(item_id) REFERENCES materials(id) ON DELETE CASCADE)"
+        val queryComments = "CREATE TABLE comments (id INTEGER PRIMARY KEY AUTOINCREMENT, sale_id INTEGER, comment TEXT, FOREIGN KEY(sale_id) REFERENCES sales(id) ON DELETE CASCADE)"
         db!!.execSQL(queryTools)
         db.execSQL(queryMaterials)
         db.execSQL(queryUsers)
         db.execSQL(querySales)
+        db.execSQL(queryComments)
     }
 
     override fun onUpgrade(db: SQLiteDatabase?, oldVersion: Int, newVersion: Int) {
-        db!!.execSQL("DROP TABLE IF EXISTS tools")
-        db.execSQL("DROP TABLE IF EXISTS materials")
-        db.execSQL("DROP TABLE IF EXISTS users")
-        db.execSQL("DROP TABLE IF EXISTS sales") // Удалите таблицу продаж при обновлении
-        onCreate(db)
+        if (oldVersion < 20) {
+            updateCommentsTable(db!!)
+        }
+        // Удаляем таблицы только при значительных изменениях схемы
+        if (oldVersion < 21) {
+            db!!.execSQL("DROP TABLE IF EXISTS tools")
+            db.execSQL("DROP TABLE IF EXISTS materials")
+            db.execSQL("DROP TABLE IF EXISTS users")
+            db.execSQL("DROP TABLE IF EXISTS sales")
+            db.execSQL("DROP TABLE IF EXISTS comments")
+            onCreate(db)
+        }
     }
 
+    private fun updateCommentsTable(db: SQLiteDatabase) {
+        db.execSQL("CREATE TABLE comments_new (id INTEGER PRIMARY KEY AUTOINCREMENT, sale_id INTEGER, comment TEXT, FOREIGN KEY(sale_id) REFERENCES sales(id) ON DELETE CASCADE)")
+        db.execSQL("INSERT INTO comments_new (sale_id, comment) SELECT sale_id, comment FROM comments")
+        db.execSQL("DROP TABLE comments")
+        db.execSQL("ALTER TABLE comments_new RENAME TO comments")
+    }
 
-    // Методы для работы с инструментами
-    // DBHelper.kt
     fun addTool(tool: Tool) {
         val values = ContentValues()
         values.put("image", tool.image)
@@ -53,22 +66,18 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
     }
 
-    // DBHelper.kt
     fun getTools(): List<Tool> {
         val tools = mutableListOf<Tool>()
         val db = this.readableDatabase
         var cursor = db.rawQuery("SELECT * FROM tools", null)
 
-        // Проверка, пуста ли таблица инструментов
         if (cursor.count == 0) {
-            // Заполнение таблицы инструментов начальными данными на русском языке
             addTool(Tool(1, "hammer", "Молоток", "Обычный молоток для общего использования", 150.99, "Ручные инструменты"))
-            addTool(Tool(2, "otvirtka", "Отвертка", "Обычная отвертка для общего использования", 49.99, "Ручные инструменты"))
+            addTool(Tool(2, "otvertka", "Отвертка", "Обычная отвертка для общего использования", 49.99, "Ручные инструменты"))
             addTool(Tool(3, "drel", "Дрель", "Мощная дрель для сверления отверстий", 4990.99, "Электроинструменты"))
             addTool(Tool(4, "pila", "Пила", "Пила для распиловки древесины", 1200.90, "Ручные инструменты"))
             addTool(Tool(5, "gaichniu_cluch", "Гаечный ключ", "Гаечный ключ для затяжки болтов", 999.99, "Ручные инструменты"))
 
-            // Получение обновленного списка инструментов
             cursor.close()
             val newCursor = db.rawQuery("SELECT * FROM tools", null)
             cursor = newCursor
@@ -92,7 +101,6 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
         return tools
     }
-
 
     fun getToolById(toolId: Int): Tool? {
         val db = this.readableDatabase
@@ -132,7 +140,6 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
     }
 
-    // Методы для работы с расходными материалами
     fun addMaterial(material: Material) {
         val values = ContentValues()
         values.put("image", material.image)
@@ -146,22 +153,18 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
     }
 
-    // DBHelper.kt
     fun getMaterials(): List<Material> {
         val materials = mutableListOf<Material>()
         val db = this.readableDatabase
         var cursor = db.rawQuery("SELECT * FROM materials", null)
 
-        // Проверка, пуста ли таблица материалов
         if (cursor.count == 0) {
-            // Заполнение таблицы материалов начальными данными на русском языке
             addMaterial(Material(1, "gvozdi", "Гвозди", "Гвозди для крепления", 1.0, "Крепежные материалы"))
             addMaterial(Material(2, "vinti", "Винты M6", "Винты для крепления", 2.0, "Крепежные материалы"))
             addMaterial(Material(3, "bolti", "Болты M6", "Болты для крепления", 3.0, "Крепежные материалы"))
-            addMaterial(Material(4, "thaibi", "Шайбы M6", "Шайбы для крепления", 4.0, "Крепежные материалы"))
+            addMaterial(Material(4, "shaibi", "Шайбы M6", "Шайбы для крепления", 4.0, "Крепежные материалы"))
             addMaterial(Material(5, "gaiki", "Гайки M6", "Гайки для крепления", 5.0, "Крепежные материалы"))
 
-            // Получение обновленного списка материалов
             cursor.close()
             val newCursor = db.rawQuery("SELECT * FROM materials", null)
             cursor = newCursor
@@ -185,7 +188,6 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
         return materials
     }
-
 
     fun getMaterialById(materialId: Int): Material? {
         val db = this.readableDatabase
@@ -225,7 +227,6 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
     }
 
-    // Методы для работы с пользователями
     fun addUser(user: User) {
         val values = ContentValues()
         values.put("email", user.email)
@@ -267,7 +268,6 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         return user
     }
 
-    // DBHelper.kt
     fun getUserByEmail(email: String): User? {
         val db = this.readableDatabase
         val cursor = db.rawQuery("SELECT * FROM users WHERE email = '$email'", null)
@@ -288,10 +288,6 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.close()
         return user
     }
-
-
-
-
 
     fun getUserIdByEmail(email: String): Int {
         val db = this.readableDatabase
@@ -345,17 +341,23 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         db.update("users", values, "id = ?", arrayOf(userId.toString()))
         db.close()
     }
-    // DBHelper.kt
-    fun addSale(itemId: Int, itemType: String, quantity: Int, saleDate: String) {
-        val values = ContentValues()
-        values.put("item_id", itemId)
-        values.put("item_type", itemType)
-        values.put("quantity", quantity)
-        values.put("sale_date", saleDate)
+
+    fun addSale(itemId: Int, itemType: String, itemName: String, itemImage: String, itemDescription: String, price: Double, quantity: Int, saleDate: String): Int {
+        val values = ContentValues().apply {
+            put("item_id", itemId)
+            put("item_type", itemType)
+            put("item_name", itemName)
+            put("item_image", itemImage)
+            put("item_description", itemDescription)
+            put("price", price)
+            put("quantity", quantity)
+            put("sale_date", saleDate)
+        }
 
         val db = this.writableDatabase
-        db.insert("sales", null, values)
+        val saleId = db.insert("sales", null, values).toInt()
         db.close()
+        return saleId
     }
 
     fun getSales(): List<Sale> {
@@ -369,6 +371,10 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
                     id = cursor.getInt(cursor.getColumnIndexOrThrow("id")),
                     itemId = cursor.getInt(cursor.getColumnIndexOrThrow("item_id")),
                     itemType = cursor.getString(cursor.getColumnIndexOrThrow("item_type")),
+                    itemName = cursor.getString(cursor.getColumnIndexOrThrow("item_name")),
+                    itemImage = cursor.getString(cursor.getColumnIndexOrThrow("item_image")),
+                    itemDescription = cursor.getString(cursor.getColumnIndexOrThrow("item_description")),
+                    price = cursor.getDouble(cursor.getColumnIndexOrThrow("price")),
                     quantity = cursor.getInt(cursor.getColumnIndexOrThrow("quantity")),
                     saleDate = cursor.getString(cursor.getColumnIndexOrThrow("sale_date"))
                 )
@@ -379,6 +385,46 @@ class DBHelper(val context: Context, factory: SQLiteDatabase.CursorFactory?) :
         cursor.close()
         db.close()
         return sales
+    }
+
+    fun addComment(saleId: Int, comment: String) {
+        val db = this.writableDatabase
+        val values = ContentValues().apply {
+            put("sale_id", saleId)
+            put("comment", comment)
+        }
+        db.insert("comments", null, values)
+        db.close()
+    }
+
+    fun getCommentsBySaleId(saleId: Int): List<Comment> {
+        val db = this.readableDatabase
+        val query = "SELECT * FROM comments WHERE sale_id = ?"
+        val result = db.rawQuery(query, arrayOf(saleId.toString()))
+        val comments = mutableListOf<Comment>()
+
+        val idIndex = result.getColumnIndex("id")
+        val saleIdIndex = result.getColumnIndex("sale_id")
+        val commentIndex = result.getColumnIndex("comment")
+
+        if (idIndex != -1 && saleIdIndex != -1 && commentIndex != -1) {
+            if (result.moveToFirst()) {
+                do {
+                    val comment = Comment(
+                        result.getInt(idIndex),
+                        result.getInt(saleIdIndex),
+                        result.getString(commentIndex)
+                    )
+                    comments.add(comment)
+                } while (result.moveToNext())
+            }
+        } else {
+            Log.e("DBHelper", "One or more columns not found in comments table")
+        }
+
+        result.close()
+        db.close()
+        return comments
     }
 
 }
